refactor: split group and entry ui
